fix: Add Title to Storybook
