+++ src/app/components/atoms/Title/title.tsx
@@ -1,0 +1,22 @@
+interface TitleProps {
+  children: React.ReactNode,
+  classes?: string,
+  size?: string,
+}
+
+export default function Title({children, classes, size}: TitleProps) {
+  if (size === undefined){
+    size = "text-xl md:text-3xl"
+  }
+  return (
+    <div className='flex flex-col justify-center items-center'>
+      <div className={`
+        font-bold text-green-primary font-rufina
+        ${size} ${classes}
+      `}>
+        {children}
+      </div>
+    </div>
+  )
+}
+
